update and refactor ActionDetailRecord
