+++ packages/modules/token/src/actions/on-chain/deployInfinitERC20Burnable.test.ts
@@ -1,17 +1,17 @@
-import { beforeAll, describe, expect, test, vi } from 'vitest'
+import { beforeAll, describe, expect, test } from 'vitest'
 
 import { encodeFunctionData } from 'viem'
 import { Address, privateKeyToAccount } from 'viem/accounts'
 
 import { ANVIL_PRIVATE_KEY } from '@actions/__mocks__/account'
 import { TEST_ADDRESSES } from '@actions/__mocks__/address'
-import { DeployInfinitERC20Action } from '@actions/deployInfinitERC20'
+import { DeployInfinitERC20BurnableAction } from '@actions/on-chain/deployInfinitERC20Burnable'
 
 import { TestChain, TestInfinitWallet } from '@infinit-xyz/test'
 import { readArtifact } from '@utils/artifact'
 
-describe('deployInfinitERC20Action', () => {
-  let action: DeployInfinitERC20Action
+describe('deployInfinitERC20BurnableBurnableAction', () => {
+  let action: DeployInfinitERC20BurnableAction
   let client: TestInfinitWallet
   let bobClient: TestInfinitWallet
 
@@ -26,7 +26,7 @@ describe('deployInfinitERC20Action', () => {
   })
 
   test('deploy token, owner is deployer', async () => {
-    action = new DeployInfinitERC20Action({
+    action = new DeployInfinitERC20BurnableAction({
       params: {
         owner: client.walletClient.account.address,
         name: 'TestToken',
@@ -45,11 +45,10 @@ describe('deployInfinitERC20Action', () => {
     expect(curRegistry.tokens).not.toBe(null)
     // get all values of curRegistry.tokens
     const tokens: Address[] = Object.keys(curRegistry.tokens!) as Address[]
-    // for (const token of tokens) {
     const ttToken = tokens[0]
-    expect(curRegistry.tokens![tokens[0]].type).toBe('InfinitERC20')
+    expect(curRegistry.tokens![tokens[0]].type).toBe('InfinitERC20Burnable')
 
-    const infinitERC20 = await readArtifact('InfinitERC20')
+    const infinitERC20 = await readArtifact('InfinitERC20Burnable')
     const ownerBal = await client.publicClient.readContract({
       address: ttToken,
       abi: infinitERC20.abi,
@@ -58,7 +57,6 @@ describe('deployInfinitERC20Action', () => {
     })
     expect(ownerBal).toBe(BigInt(500000000000000000000000))
 
-    // test totalsupply
     const totalSupply = await client.publicClient.readContract({
       address: ttToken,
       abi: infinitERC20.abi,
@@ -109,7 +107,7 @@ describe('deployInfinitERC20Action', () => {
 
   // test deploy token, owner is not deployer
   test('deploy token, owner is not deployer', async () => {
-    action = new DeployInfinitERC20Action({
+    action = new DeployInfinitERC20BurnableAction({
       params: {
         owner: bob,
         name: 'TestToken',
@@ -128,11 +126,10 @@ describe('deployInfinitERC20Action', () => {
     expect(curRegistry.tokens).not.toBe(null)
     // get all values of curRegistry.tokens
     const tokens: Address[] = Object.keys(curRegistry.tokens!) as Address[]
-    // for (const token of tokens) {
     const ttToken = tokens[0]
-    expect(curRegistry.tokens![tokens[0]].type).toBe('InfinitERC20')
+    expect(curRegistry.tokens![tokens[0]].type).toBe('InfinitERC20Burnable')
 
-    const infinitERC20 = await readArtifact('InfinitERC20')
+    const infinitERC20 = await readArtifact('InfinitERC20Burnable')
     const deployerBal = await client.publicClient.readContract({
       address: ttToken,
       abi: infinitERC20.abi,
@@ -149,7 +146,6 @@ describe('deployInfinitERC20Action', () => {
     })
     expect(ownerBal).toBe(BigInt(500000000000000000000000))
 
-    // test totalsupply
     const totalSupply = await client.publicClient.readContract({
       address: ttToken,
       abi: infinitERC20.abi,
@@ -169,7 +165,7 @@ describe('deployInfinitERC20Action', () => {
 
   // test deploy token, not owner cant mint
   test('deploy token, not owner cant mint', async () => {
-    action = new DeployInfinitERC20Action({
+    action = new DeployInfinitERC20BurnableAction({
       params: {
         owner: client.walletClient.account.address,
         name: 'TestToken',
@@ -186,7 +182,7 @@ describe('deployInfinitERC20Action', () => {
     const tokens: Address[] = Object.keys(curRegistry.tokens!) as Address[]
     // owner mint tokens
     const ttToken = tokens[0]
-    const infinitERC20 = await readArtifact('InfinitERC20')
+    const infinitERC20 = await readArtifact('InfinitERC20Burnable')
 
     const mintAmount = BigInt(100000000000000000000000)
 
@@ -212,24 +208,54 @@ describe('deployInfinitERC20Action', () => {
     ).rejects.toThrowError()
   })
 
-  test('deploy token, test initial supply > max supply', async () => {
-    vi.spyOn(console, 'error').mockImplementationOnce(() => {})
-    try {
-      action = new DeployInfinitERC20Action({
-        params: {
-          owner: client.walletClient.account.address,
-          name: 'TestToken',
-          symbol: 'TT',
-          maxSupply: BigInt(1000000000000000000000000000),
-          initialSupply: BigInt(2000000000000000000000000000),
+  // test deploy token, test burn token
+  test('deploy token, burn token', async () => {
+    action = new DeployInfinitERC20BurnableAction({
+      params: {
+        owner: client.walletClient.account.address,
+        name: 'TestToken',
+        symbol: 'TT',
+        maxSupply: BigInt(1000000000000000000000000000),
+        initialSupply: BigInt(500000000000000000000000),
+      },
+      signer: {
+        deployer: client,
+      },
+    })
+    const curRegistry = await action.run({}, undefined, undefined)
+
+    const tokens: Address[] = Object.keys(curRegistry.tokens!) as Address[]
+    // owner mint tokens
+    const ttToken = tokens[0]
+    const infinitERC20 = await readArtifact('InfinitERC20Burnable')
+
+    const burnAmount = BigInt(100000000000000000000000)
+    const balanceBeforeBurn = await client.publicClient.readContract({
+      address: ttToken,
+      abi: infinitERC20.abi,
+      functionName: 'balanceOf',
+      args: [client.walletClient.account.address],
+    })
+
+    await client.sendTransactions([
+      {
+        name: 'burn',
+        txData: {
+          to: ttToken,
+          data: encodeFunctionData({
+            abi: infinitERC20.abi,
+            functionName: 'burn',
+            args: [burnAmount],
+          }),
         },
-        signer: {
-          deployer: client,
-        },
-      })
-      await action.run({}, undefined, undefined)
-    } catch (e: any) {
-      expect(e.details.includes('9e79f854'))
-    }
+      },
+    ])
+    const balanceAfterBurn = await client.publicClient.readContract({
+      address: ttToken,
+      abi: infinitERC20.abi,
+      functionName: 'balanceOf',
+      args: [client.walletClient.account.address],
+    })
+    expect(balanceAfterBurn).toBe(balanceBeforeBurn - burnAmount)
   })
 })
