+++ packages/modules/token/src/index.ts
@@ -1,3 +1,2 @@
 export { TokenRegistry as Registry } from '@/src/type'
 export { actions } from '@actions/index'
-export { offChainActions } from '@offChainActions/index'
