--- packages/modules/token/src/off-chain-actions/index.ts
@@ -1,13 +1,0 @@
-import { InfinitOffChainActionRecord } from '@infinit-xyz/core'
-
-import { GetMerkleTreeProofsOffChainAction, GetMerkleTreeProofsOffChainActionParamsSchema } from '@offChainActions/getMerkleTreeProofs'
-
-export const offChainActions = {
-  getProofMerkleTree: {
-    name: 'Get Merkle Tree Proofs',
-    offChainActionClassName: GetMerkleTreeProofsOffChainAction.name,
-    paramsSchema: GetMerkleTreeProofsOffChainActionParamsSchema,
-  },
-} satisfies InfinitOffChainActionRecord
-
-export { GetMerkleTreeProofsOffChainAction }
